EightMilesAPI/EightMilesAPI/requirements.txt
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.6
pip==10.0.1
PyJWT==1.4.2
PyMySQL==0.9.3
setuptools==39.0.1
SQLAlchemy==1.3.13
Werkzeug==0.16.1
Flask-Cors==3.0.8